email/requirements.txt
aiven-client==4.7.3
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
debugpy==1.8.13
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
Faker==37.0.2
idna==3.10
kafka-python==2.2.11
python-decouple==3.8
requests==2.32.4
requests-toolbelt==1.0.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.4.0
